Block: move "fromRPC" functions into constructors
