+++ packages/block/src/block/index.ts
@@ -1,3 +1,2 @@
 export * from './block.js'
 export * from './constructors.js'
-export * from './from-rpc.js'
