+++ packages/block/src/header/index.ts
@@ -1,3 +1,2 @@
 export * from './constructors.js'
 export * from './header.js'
-export * from './header-from-rpc.js'
